test_modules: Netperf.py: fix incorrect deviation calculation
int conversion of result["confidence"] was truncating everything after
decimal point, therefore division was incorrect for numbers with
non-zero decimal part.

+++ test_modules/Netperf.py
@@ -519,7 +519,7 @@ class Netperf(TestGeneric):
             rate_deviation = 2*res_data["std_deviation"]
         elif len(rates) == 1 and self._confidence is not None:
             result = results[0]
-            rate_deviation = rate * (int(result["confidence"][1]) / 100)
+            rate_deviation = rate * (float(result["confidence"][1]) / 100)
         else:
             rate_deviation = 0.0
